fix /W1 warnings

+++ src/pf_imgui/elements/Combobox.h
@@ -129,7 +129,7 @@ class PF_IMGUI_EXPORT Combobox : public CustomCombobox<T, Selectable>,
     assert(index < items.size());
     if (index != selectedItemIndex) {
       if (selectedItemIndex.has_value()) { filteredItems[*selectedItemIndex]->second->setValue(false); }
-      selectedItemIndex = index;
+      selectedItemIndex = static_cast<int>(index);
       filteredItems[*selectedItemIndex]->second->setValue(true);
       ValueObservable<T>::setValueInner(filteredItems[index]->first);
       ValueObservable<T>::notifyValueChanged();
@@ -148,7 +148,7 @@ class PF_IMGUI_EXPORT Combobox : public CustomCombobox<T, Selectable>,
       const auto selectedItemAsString = *src["selected"].value<std::string>();
       for (const auto &[idx, item] : items | ranges::views::enumerate) {
         if (item.second->getLabel() == selectedItemAsString) {
-          selectedItemIndex = idx;
+          selectedItemIndex = static_cast<int>(idx);
           items[idx].second->setValue(true);
           ValueObservable<T>::setValueAndNotifyIfChanged(items[idx].first);
           break;
